Python: Support add column
